fix(warning): clean up redundant toFloat conversions in NoTuneAmbientCanvas

+++ app/src/main/kotlin/com/music/echo/notune/theme/NoTuneAmbientCanvas.kt
@@ -108,7 +108,7 @@ fun NoTuneAmbientCanvas(
             drawCircle(
                 color = if (i % 2 == 0) accentColor.copy(alpha = particleAlpha) else secondaryColor.copy(alpha = particleAlpha),
                 radius = (3 + (i % 4) * 2).dp.toPx(),
-                center = Offset(px.toFloat(), py.toFloat())
+                center = Offset(px, py)
             )
         }
